LLM/website-rag-assistant/requirements.txt
openai
faiss-cpu
numpy
flask
requests
python-dotenv
beautifulsoup4
lxml
tiktoken